requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==7.1.2
click-didyoumean==0.3.1
contourpy==1.3.2
croniter==6.0.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
fonttools==4.61.0
frozenlist==1.8.0
google-auth==2.47.0
google-genai==1.57.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.6
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
marshmallow==3.26.2
matplotlib==3.10.7
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.1
p==1.5.0
packaging==25.0
pillow==12.0.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.2.5
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
redis==7.1.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.6.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
